requirements.txt
langchain
chromadb
openai
tiktoken
python-dotenv
streamlit
pandas
google-generativeai
requests